+++ lib/uo/location.go
@@ -3,11 +3,11 @@ package uo
 // Location identifies an absolute location in the universe.
 type Location struct {
 	// Absolute X position on the map [0-)
-	X int
+	X int16
 	// Absolute Y position on the map
-	Y int
+	Y int16
 	// Absolute Z position on the map
-	Z int
+	Z int8
 }
 
 // This Location value indicates to a container that the item should be placed
@@ -18,16 +18,16 @@ var RandomContainerLocation Location = Location{X: RandomDropX, Y: RandomDropY}
 // map.
 func (l Location) WrapToOverworld() Location {
 	for l.X < 0 {
-		l.X += MapOverworldWidth
+		l.X += int16(MapOverworldWidth)
 	}
-	for l.X >= MapOverworldWidth {
-		l.X -= MapOverworldWidth
+	for l.X >= int16(MapOverworldWidth) {
+		l.X -= int16(MapOverworldWidth)
 	}
 	for l.Y < 0 {
-		l.Y += MapHeight
+		l.Y += int16(MapHeight)
 	}
-	for l.Y >= MapHeight {
-		l.Y -= MapHeight
+	for l.Y >= int16(MapHeight) {
+		l.Y -= int16(MapHeight)
 	}
 	return l
 }
@@ -35,17 +35,17 @@ func (l Location) WrapToOverworld() Location {
 // WrapToDungeonServer returns the location wrapped to the dungeon server
 // section of the map.
 func (l Location) WrapToDungeonServer() Location {
-	for l.X < MapOverworldWidth {
-		l.X += MapWidth - MapOverworldWidth
+	for l.X < int16(MapOverworldWidth) {
+		l.X += int16(MapWidth) - int16(MapOverworldWidth)
 	}
-	for l.X > MapWidth {
-		l.X -= MapWidth - MapOverworldWidth
+	for l.X > int16(MapWidth) {
+		l.X -= int16(MapWidth) - int16(MapOverworldWidth)
 	}
 	for l.Y < 0 {
-		l.Y += MapHeight
+		l.Y += int16(MapHeight)
 	}
-	for l.Y >= MapHeight {
-		l.Y -= MapHeight
+	for l.Y >= int16(MapHeight) {
+		l.Y -= int16(MapHeight)
 	}
 	return l
 }
@@ -55,7 +55,7 @@ func (l Location) WrapToDungeonServer() Location {
 // wrapping as appropriate based on the reference location.
 func (l Location) WrapAndBound(ref Location) Location {
 	ref = ref.Bound()
-	if ref.X < MapOverworldWidth {
+	if ref.X < int16(MapOverworldWidth) {
 		return l.WrapToOverworld()
 	} else {
 		return l.WrapToDungeonServer()
@@ -66,16 +66,16 @@ func (l Location) WrapAndBound(ref Location) Location {
 // of the map.
 func (l Location) Bound() Location {
 	for l.X < 0 {
-		l.X += MapWidth
+		l.X += int16(MapWidth)
 	}
-	for l.X > MapWidth {
-		l.X -= MapWidth
+	for l.X > int16(MapWidth) {
+		l.X -= int16(MapWidth)
 	}
 	for l.Y < 0 {
-		l.Y += MapHeight
+		l.Y += int16(MapHeight)
 	}
-	for l.Y >= MapHeight {
-		l.Y -= MapHeight
+	for l.Y >= int16(MapHeight) {
+		l.Y -= int16(MapHeight)
 	}
 	if l.Z < MapMinZ {
 		l.Z = MapMinZ
@@ -88,7 +88,7 @@ func (l Location) Bound() Location {
 
 // XYDistance returns the maximum distance from l to d along either the X or Y
 // axis.
-func (l Location) XYDistance(d Location) int {
+func (l Location) XYDistance(d Location) int16 {
 	dx := l.X - d.X
 	dy := l.Y - d.Y
 	if dx < 0 {
